move GreedyPathConverter
