+++ localstack-core/localstack/services/apigateway/next_gen/execute_api/handlers/resource_router.py
@@ -13,8 +13,8 @@ from localstack.aws.protocol.routing import (
     transform_path_params_to_rule_vars,
 )
 from localstack.http import Response
+from localstack.http.router import GreedyPathConverter
 from localstack.services.apigateway.models import RestApiDeployment
-from localstack.services.edge import GreedyPathConverter
 
 from ..api import RestApiGatewayHandler, RestApiGatewayHandlerChain
 from ..context import RestApiInvocationContext
