+++ tests/unit/aws/protocol/test_op_router.py
@@ -5,7 +5,7 @@ from werkzeug.routing import Map, Rule
 from localstack.aws.protocol.op_router import RestServiceOperationRouter
 from localstack.aws.spec import list_services, load_service
 from localstack.http import Request
-from localstack.services.edge import GreedyPathConverter
+from localstack.http.router import GreedyPathConverter
 
 
 def _collect_services():
